Update some number usage

+++ src/events/event.rs
@@ -4,6 +4,6 @@ use super::event_data::EventData;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Event {
-    pub version: u64,
+    pub version: u32,
     pub data: EventData,
 }
