+++ src/projections/has_listen.rs
@@ -1,4 +1,4 @@
 pub trait HasListen {
     fn has_listen(&self, artist_name: &str, track_name: &str, end_time: &str) -> bool;
-    fn version(&self) -> u64;
+    fn version(&self) -> u32;
 }
